+++ source/TeleEngine.NET/Components/Vertices/DefaultModels/Models/UnknownModel.cs
@@ -6,7 +6,7 @@ namespace TeleEngine.NET.Components.Vertices.DefaultModels.Models
         public override VertexModel Model => new()
         {
             Vertices = new float[] { 1 },
-            Indexes = new uint[] { 1 }
+            Indexes = new float[] { 1 }
         };
     }
 }
